Move PlanetaryOrbit to its own file
